Rename blacklisted peers variables
