+++ test/integration/scenarios/index.js
@@ -18,7 +18,7 @@ module.exports = {
 	network: {
 		peers: require('./network/peers'),
 		peerDisconnect: require('./network/peer_disconnect'),
-		peerBlacklist: require('./network/peer_blacklist'),
+		peersBlackList: require('./network/peers_blacklist'),
 	},
 	propagation: {
 		blocks: require('./propagation/blocks'),
