+++ test/integration/scenarios/network/peers_blacklist.js
