+++ test/unit/modules/system.js
@@ -48,7 +48,7 @@ describe('system', () => {
 			nethash: 1,
 			minVersion: '>=1.0.0-beta.0',
 			nonce: 1,
-			blackListPeers: [],
+			blackListedPeers: [],
 		};
 
 		systemModule = new System(
